Make it compatible without the TurboModule
