+++ android/build.gradle
@@ -49,6 +49,15 @@ android {
     sourceCompatibility JavaVersion.VERSION_1_8
     targetCompatibility JavaVersion.VERSION_1_8
   }
+  sourceSets {
+    main {
+      if (isNewArchitectureEnabled()) {
+          java.srcDirs += ['src/turbo']
+      } else {
+          java.srcDirs += ['src/legacy']
+      }
+    }
+  }
 }
 
 repositories {
