+++ android/src/main/java/com/reactnativeimageresizer/ImageResizerModuleImpl.java
@@ -5,11 +5,8 @@ import android.content.Context;
 import android.graphics.Bitmap;
 import android.net.Uri;
 import android.os.AsyncTask;
-import android.os.Handler;
-import android.util.Base64;
 import android.util.Log;
 
-import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.facebook.react.bridge.Arguments;
@@ -18,41 +15,25 @@ import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.WritableMap;
-import com.facebook.react.module.annotations.ReactModule;
 
-import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
 import java.util.UUID;
 
-@ReactModule(name = ImageResizerModule.NAME)
-public class ImageResizerModule extends NativeImageResizerSpec {
-    public static final String NAME = "ImageResizer";
-    private Context context;
+public class ImageResizerModuleImpl {
+  public static final String NAME = "ImageResizer";
 
-    public ImageResizerModule(ReactApplicationContext reactContext) {
-      super(reactContext);
-      this.context = reactContext;
-    }
-
-  @Override
-  @NonNull
-  public String getName() {
-      return NAME;
-  }
-
-  @Override
-  public void createdResizedImage(String uri, double width, double height, String format, double quality, Double rotation, @Nullable String outputPath, Boolean keepMeta, @Nullable String mode, Boolean onlyScaleDown, Promise promise) {
+  public static void createdResizedImage(String uri, double width, double height, String format, double quality, Double rotation, @Nullable String outputPath, Boolean keepMeta, @Nullable String mode, Boolean onlyScaleDown, Promise promise, Context context, ReactApplicationContext reactContext) {
     WritableMap options = Arguments.createMap();
     options.putString("mode", mode);
     options.putBoolean("onlyScaleDown", onlyScaleDown);
 
     // Run in guarded async task to prevent blocking the React bridge
-    new GuardedAsyncTask<Void, Void>(getReactApplicationContext()) {
+    new GuardedAsyncTask<Void, Void>(reactContext) {
       @Override
       protected void doInBackgroundGuarded(Void... params) {
         try {
-          Object response = createResizedImageWithExceptions(uri, (int) width, (int) height, format, (int) quality, rotation.intValue(), outputPath, keepMeta, options);
+          Object response = createResizedImageWithExceptions(uri, (int) width, (int) height, format, (int) quality, rotation.intValue(), outputPath, keepMeta, options, context);
           promise.resolve(response);
         }
         catch (IOException e) {
@@ -63,15 +44,15 @@ public class ImageResizerModule extends NativeImageResizerSpec {
   }
 
   @SuppressLint("LongLogTag")
-  private Object createResizedImageWithExceptions(String imagePath, int newWidth, int newHeight,
+  private static Object createResizedImageWithExceptions(String imagePath, int newWidth, int newHeight,
                                                   String compressFormatString, int quality, int rotation, String outputPath,
                                                   final boolean keepMeta,
-                                                  final ReadableMap options) throws IOException {
+                                                  final ReadableMap options, Context context) throws IOException {
 
     Bitmap.CompressFormat compressFormat = Bitmap.CompressFormat.valueOf(compressFormatString);
     Uri imageUri = Uri.parse(imagePath);
 
-    Bitmap scaledImage = ImageResizer.createResizedImage(this.context, imageUri, newWidth, newHeight, quality, rotation,
+    Bitmap scaledImage = ImageResizer.createResizedImage(context, imageUri, newWidth, newHeight, quality, rotation,
       options.getString("mode"), options.getBoolean("onlyScaleDown"));
 
     if (scaledImage == null) {
@@ -89,23 +70,17 @@ public class ImageResizerModule extends NativeImageResizerSpec {
 
     // If resizedImagePath is empty and this wasn't caught earlier, throw.
     if (resizedImage.isFile()) {
-      ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
-      scaledImage.compress(Bitmap.CompressFormat.PNG, 100, byteArrayOutputStream);
-      byte[] byteArray = byteArrayOutputStream .toByteArray();
-      String base64 = Base64.encodeToString(byteArray, Base64.DEFAULT);
-
       response.putString("path", resizedImage.getAbsolutePath());
       response.putString("uri", Uri.fromFile(resizedImage).toString());
       response.putString("name", resizedImage.getName());
       response.putDouble("size", resizedImage.length());
       response.putDouble("width", scaledImage.getWidth());
       response.putDouble("height", scaledImage.getHeight());
-      response.putString("base64", base64);
 
       // Copy file's metadata/exif info if required
       if(keepMeta){
         try{
-          ImageResizer.copyExif(this.context, imageUri, resizedImage.getAbsolutePath());
+          ImageResizer.copyExif(context, imageUri, resizedImage.getAbsolutePath());
         }
         catch(Exception ignored){
           Log.e("ImageResizer::createResizedImageWithExceptions", "EXIF copy failed", ignored);
